requirements.txt
sentence-transformers
faiss-cpu
streamlit
pypdf
python-docx
apscheduler
python-dotenv
openai
